Add user properties, storing information for user authentication
The user's unique ID string associated with a single IdP is currently stored in the User entity as certificateDn. Instead, a UserProperty entity can allow each user to be associated with one or more IdPs. This will allow account linking across multiple IdPs, as well as flexibility to add IdPs without further changes to the database.