Create entity to store account linking requests
The LinkAccountRequest entity has a similar role to the RetrieveAccountRequest entity, storing the user that new log-in details are to be added to, as well as the required confirmation code and the user's id string. Unlike RetrieveAccountRequest, this entity may also store the user that created the request (as they may or may not be registered), as well as the authentication type of the log-in to be added.