Add user properties on registration
When a new user registers, their details are now added as a user property rather than being stored in certificateDn. certificateDn is currently overwritten with the user id after creation of the user to ensure the id has been created, and so the value is non-null and unique. An alternative placeholder may be used later.