Delete user account on successful linking
If an account has been successfully linked to another account, this account must be deleted to prevent the id string being associated with two different accounts. The function to add user properties has also been modified to prevent overwriting by default. It is useful to be sure if the intention is for a property to be overwritten rather than added, and when attempting to overwrite there currently appears to be an uncaught error anyway.