Reimplement account recovery via account linking
These changes to the account linking script/entity allow account recovery too. The two processes are almost identical, so it is convenient to test the implementation using the same structures. If the entities are later seperated, the queries used may be generalised, or the scripts may also be seperated, but will likely share a number of functions.