Fix account linking if no user properties
If the "primary" user does not have user properties, their certificateDn is now added as a new property with the auth type set using the linking request. Previously in this case certificateDn was overwritten and the id string was not saved. Overwriting the certificateDn is still carried out, although only if necessary, as if the user has properties this will already have been performed.