Store auth types of both login methods
Modified the LinkAccountRequest entity to store the "primary" and "secondary" auth type. The primary auth type is entered by the user (although in most cases should be known via user properties from the primary account). The secondary auth type is known as the user is authenticated with this method when making the request, regardless of whether they are registered. The expectation of secondaryUser being a \User on construction has also been removed, as this led to errors for null (unregistered) users.