Undo changes to account retrieval
Account retrieval/recovery is now (implicitly) carried out via account linking. These files are no longer in use, but have been reverted to the original workflow to be self consistent, as they currently make use of the account linking process, but will not be updated as changes to this process are made.