Request roles from account when linking
When linking two registered accounts, self-revoke all roles associated with the account to be deleted, and request these roles for the account that remains for autditability/to intervene less with entity roles. Also added a check to remove linking requests where the account to be deleted is the "primary" account, in addition to the "secondary" account in a request, as both cases could lead to errors.