Prevent account linking to same account
For account recovery, no check is done between the current id string and the id string to be recovered, as if they are the same then the user would be logged in and recovery would not be an option. For account linking, both id strings will be valid and may be registered, so it must be checked that the user does not attempt to link to the account they are already using.