Update controllers and views for account recovery
The new account recovery process, which will allow both users to be registered, handle user properties and request transfer of user roles, is almost identical to the account linking process. This new process will therefore (initially) be implemented using the same script, and so the controllers and views for this process have been updated accordingly.