Change onchange to onclick for auth types to allow same option to be reselected
On navigation to the link account page, options selected are retained, but not recognised as selected. Validation of the auth type and enabling of the id string input were tied to the auth type changing, but this requires the user to change from the currently selected option and back if they have navigated to the page. Using onclick, they can simply reselect the same option, which is still not ideal, but is an improvement.