Merge roles on linking
Rather than requesting all roles, instead grant roles owned by the user making the linking request to the account being linked to. The account being linked to must exist, whereas when making a request, the user does not need to be registered, so the account being linked to (and its roles) are preserved. But, if the user has roles with the account making the request, it is unnecessary to revoke and re-request these roles, as they had been granted to that same user. This also avoids needing to check if a user exists to approve the roles. Also changed duplicate roles to continue to next role rather than raise an exception.