Validate input for account linking
Validation is performed on the inputted auth type, id string and email, using regex and checking an input has been entered. Different validation is performed and messages are output depending on whether the selected auth type is the same as the type currently in use. If the methods are the same, the user will perform account recovery, resulting in the id entered being overwritten, so a warning is necessary. A stronger warning should be given if attempting this for non-certificate methods, but currently the opposite is implemented for testing.