Add user input for auth type
The "primary" auth type is not always known, as the user may not have logged in recently to add user properties. This allows the user to enter the auth type manually. Validation of the id string will be done based on this input, and if the input matches the current auth type, a warning message will display, as this will lead to account "recovery" rather than "linking", overwriting the old id string.