Request and attempt to grant roles on account linking
Previously any roles that the account to be deleted had that the primary account did not were granted automatically. Instead, the primary account automatically requests these roles, and then attempts to grant itself them. The other account then also attempts to grant the roles, and if this fails the role remains pending. The other account also revokes all its roles before it is deleted. This provides stronger records of changes being made to role permissions. This also results in no changes being made that could not be approved by the user linking their accounts, while removing the complication of having to manually go through that process.