Prevent user from registering if id string stored as certificateDn
Before adding user properties, on registration it was checked that no user was registered with the same certificateDn. When user properties were added, getUserByPrinciple was updated to check properties instead of certificateDn, as this is where the id strings should now be stored. However, if the user has not yet logged in since this change, their id string will still be stored as a certificateDn, so when a user attempts to register, this must be checked in addition to checking user properties. Note that in either case, the error should be caught before reaching this point, but it is a useful additional check.