Allow non-admin users to view Scopes page
The scope/scopes views and controllers have been moved out of their respective admin directories, as they are forming the basis for the non-admin scripts. Rather than throwing an error if the user is not an admin, the controllers instead pass a flag to the views, hiding admin controls, although the controls would not work for a non-admin user anyway. References to Admin_Scopes pages have been updated to Scopes, and the page has been added to the menu for non-admin users. The Scope_Help link has also been renamed to more clearly differentiate it, as previously the Admin_Scopes and Scope_Help pages were both labelled "Scopes".