Create function for start time validation
When editing downtimes, the same check is performed on the old start time and new start time to ensure both are at least 24 hours in the future. This check will also be used to prevent downtimes within 24 hours from being deleted, so this will be performed by the new separate function.