Factor out start time validation when editing downtimes
Validation to check that a downtime's start time is not within 24 hours is performed on both the old start time and the new start time when editing a downtime. This check is the same as the one performed when attempting to delete a downtime, so the same function is now used to carry out these checks.