Update account linking/registration for null certificateDns
certificateDns are null by default, rather than required to be unique, so it no longer needs to be set to the user ID. When adding properties, certificateDn is set to null if it is not already. An additional check if the certificateDn is not null may be useful, to confirm the new property matches, or that editing is intentional.