Order auth types based on MyConfig1
Rather than hardcoding the entire auth types list, only the auth types for each token type must be hardcoded now. If the order of tokens is changed via MyConfig1, this will be reflected in the order of the authe types. In the future the auth realms may be extracted automatically too. Also added a flag to only return the "main" in-use auth types i.e. only EGI for Shibboleth, as this sublist is what will be used for API outputs and linking, but in general the other types should be valid.