Use auth types list when admin editing a user property
Previously the auth type was entered manually by the admin, but only certain types should be valid, and currently this is checked anyway, so it is more convenient to select from the list of valid types. This list is currently all possible types, rather than just the "main" types offered when linking, as this may be useful for admins. There is also currently no validation of the ID string based on the type, but this could be added in a similar way to the linking validation.