Fix set admin function
Function previously and seemingly unintentionally only allows admins to demote other admins, rather than promote new admins. The comment above the checkUserIsAdmin and error message both suggest it is checking the current user's admin status, rather than the user being changed, so this change reflects this. If this function, and the calling controller/view are not intented to be used to promote admins, this can be caught earlier and a clearer message should be given.