Update deploy scripts for user properties
Updated AddUsers to add the certificate DNs as user properties rather than in the certificateDn field. Currently all ID strings are certificates, which is hardcoded, but this could be updated if other auth types were added to the XML file. Also updated the AddXRoles scripts to query user properties rather than certificateDns.