Add functions to add properties to users
These functions allow property key/value pairs to be added to users, containing the authentication type and associated ID string. The functions are based on similar functions used to add properties to sites and services, and will require revisiting (e.g. validation). When adding the first property, certificateDn is currently overwritten by user ID, so the value is non-null and unique. An alternative placeholder may be used later.