Look up users using user identifiers
getUserByPrinciple has been repurposed to query user identifiers rather than certificateDns when searching for a user via their ID string. This ensures identifiers are correctly used throughout the code, as in almost all cases where the ID string is known, it should be stored as an identifier. The original function has been renamed so it can be used in cases such as login where an identifier may not yet exist.