Look up users using user properties
getUserByPrinciple has been repurposed to query user properties rather than certificateDns when searching for a user via their ID string. This ensures user properties are correctly used throughout the code, as in almost all cases where the ID string is known, it should be stored as a property. The original function has been renamed so it can be used in cases such as login where a property may not yet exist.