Create entity to store identity linking requests
The LinkIdentityRequest entity has a similar role to the RetrieveAccountRequest entity, storing the user that new log-in details are to be added to, as well as the required confirmation code and the user's ID string. Unlike RetrieveAccountRequest, this entity may also store the user that created the request (as they may or may not be registered), as well as the authentication types of the identifiers, and the time the request was created.