Update controller and view for admin editing user properties
Updated functionality that allowed admins to edit user DNs to update user properties, including updating the ID string and/or auth type for a property, except the one they are currently using. If a user does not have properties, this also allows one to be added, if it matches their current certificateDn.