Set certificateDn to null by default
Change the schema to allow certificateDn to be null, rather than enforcing unique values, as unique ID strings are now handled by user properties. When adding a user property to an old user with certificateDn, the ID string added as a property must match, and certificateDn is set to null.