Add functions to delete, edit and validate user properties
Added functions which allow the properties of a user to be edited and deleted. This functionality will be used for account recovery, as well as potentially admin user editing. Also added general validation for user property names and values, as well as specific requirements when adding or editing a property, including enforcing unique ID strings.