Add functions to delete, edit and validate user identifiers
Added functions which allow the identifiers of a user to be edited and deleted. This functionality will be used for account recovery, as well as potentially admin user editing. Also added general validation for identifier names and values, as well as specific requirements when adding or editing an identifier, including enforcing unique ID strings.