Add function to get auth types, users and identifiers
Rather than relisting possible auth types in multiple places, getAuthTypes offers a consistently ordered list. The order is updated based on the token order in MyConfig1, but for tokens with multiple realms, their order is hardcoded based on the current order within the tokens. getUserByPrincipleAndType behaves similarly to getUserByPrinciple, but ensures the auth type specified matches a user identifier. This is not typically required, but can be useful to confirm. getIdentifierByIdString allows identifiers to be accessed via their ID strings.